Big update provided by Janne Hyvärinen (Case).
Changes include:

* fixed pipe encoding missing the beginning of the audio and breaking gaplessness.
* fixed the encoder to not read data beyond data chunk.
* added support for 24 bit integer and 32 bit integer and floating point input data.
* added Unicode filename support for input and output files.
* added support for WAVE_FORMAT_EXTENSIBLE file format and RIFX big endian WAVE files.
* Made encoder reject input tracks with zero samples.
* Added support for printing unicode filenames to the console.
* Fixed gapless data saving, the encoder tried to store too long numbers in padding info fields and the numbers got truncated as the storage space is only 12 bits long.
* Fixed internal resampling to produce output with correct lengths.
* Fixed encoding WAV files with huge headers.
* Fixed writing empty track when input had unsupported sample rate.
* Added more detailed error message for input file rejection if the audio specs weren't supported.

+++ hmp3/src/pub/filter2.h
@@ -1,5 +1,5 @@
 /* ***** BEGIN LICENSE BLOCK *****  
- * Source last modified: $Id: filter2.h,v 1.1 2005/07/13 17:22:24 rggammon Exp $ 
+ * Source last modified: 2024-03-16, Case
  *   
  * Portions Copyright (c) 1995-2005 RealNetworks, Inc. All Rights Reserved.  
  *       
@@ -54,7 +54,7 @@ extern "C"
     void filter2_init ( int samprate, int filter_select,
                         int monodual, FILTER2_CONTROL * fc2 );
 
-    void filter2 ( short pcm[], float *buf1, float *buf2,
+    void filter2 ( float pcm[], float *buf1, float *buf2,
                    FILTER2_CONTROL * fc2 );
 
 //----------------------------------------------
